First commit.
This is the point where I decided to commit to github.
Basic functionality is there. The navigation of photos has threaded
pre-fetching, images can be discarded and the basics og the GUI are there.

The code is still a little bit messy and may require some refactoring (I'm
still getting the hang of how to program cleanly in python). The shortcuts
should also be changeable and not so obscure and things like that.